requirements.txt
# Sector Momentum Breakout Strategy - Dependencies

# Angel One SmartAPI
smartapi-python>=1.4.0

# TOTP Generation for 2FA
pyotp>=2.8.0

# Data manipulation
pandas>=2.0.0

# HTTP requests
requests>=2.28.0

# WebSocket for real-time data
websocket-client>=1.5.0

# Build tools (optional - for creating executables)
pyinstaller>=6.0.0
